Remove `LogsLevel` enumeration from logging system
The `LogsLevel` enumeration, previously part of the
`FlowSynx.Domain.Log` namespace, has been removed. This
enumeration defined logging levels such as `None`, `Verb`,
`Dbug`, `Info`, `Warn`, `Fail`, and `Crit`. Its removal
suggests a refactor or replacement of the logging mechanism
or that it is no longer required in the codebase.

#761